key_based_loss update
