+++ flowjax/train/__init__.py
@@ -1,11 +1,7 @@
 """Utilities for training flows, fitting to samples or ysing variational inference."""
 
-from .data_fit import fit_to_data
+from .loops import fit_to_data, fit_to_key_based_loss
 from .train_utils import step
 from .variational_fit import fit_to_variational_target
 
-__all__ = [
-    "fit_to_data",
-    "fit_to_variational_target",
-    "step",
-]
+__all__ = ["fit_to_key_based_loss", "fit_to_data", "fit_to_variational_target", "step"]
